moved components used only in pokemon routes into component folder colocated in pokemon
