+++ src/app/(pokemon)/_components/ModeSwitch.tsx
