+++ src/app/(pokemon)/_components/content/AboutContent.tsx
