+++ src/app/(pokemon)/_components/content/SettingsContent.tsx
