+++ src/app/(pokemon)/_components/content/StatsContent.tsx
