+++ src/app/(pokemon)/classic/_components/PokemonFeedback.tsx
@@ -11,7 +11,7 @@ import {
   pokemonToGuessAtom,
 } from "@/atoms/GameAtoms";
 import FeedbackTile from "./FeedbackTile";
-import GameOverContent from "@/components/content/GameOverContent";
+import GameOverContent from "@/app/(pokemon)/_components/content/GameOverContent";
 
 type PokemonFeedbackProps = {
   correctAnswer: Pokemon;
