+++ src/app/(pokemon)/classic/page.tsx
@@ -1,5 +1,5 @@
-import Guesses from "@/components/Guesses";
-import ModeSwitch from "@/components/ModeSwitch";
+import Guesses from "@/app/(pokemon)/_components/Guesses";
+import ModeSwitch from "@/app/(pokemon)/_components/ModeSwitch";
 import pokedex from "@/data/pokedex.json";
 import dynamic from "next/dynamic";
 
@@ -8,9 +8,12 @@ export const metadata = {
   description: "Guess the Pokémon based on hints from previous guesses",
 };
 
-const GameWrapper = dynamic(() => import("@/components/GameWrapper"), {
-  ssr: false,
-});
+const GameWrapper = dynamic(
+  () => import("@/app/(pokemon)/_components/GameWrapper"),
+  {
+    ssr: false,
+  },
+);
 
 export default async function Classic() {
   return (
