+++ src/app/(pokemon)/whosthatpokemon/_components/DailyGame.tsx
@@ -10,7 +10,7 @@ import { defaultGuesses } from "@/constants";
 import { useAtomValue, useAtom, useSetAtom } from "jotai";
 import { useEffect } from "react";
 import ImagePanel from "./ImagePanel";
-import DailyUnavailable from "@/components/DailyUnavailable";
+import DailyUnavailable from "@/app/(pokemon)/_components/DailyUnavailable";
 import { Skeleton } from "@/components/ui/Skeleton";
 import { CalendarDate } from "@internationalized/date";
